Split nflbot into individual areas
