+++ slack-nfl-bot.rb
@@ -1,4 +1,5 @@
 require 'slack-ruby-bot'
+require 'slack-nfl-bot/database'
 require 'slack-nfl-bot/version'
 require 'slack-nfl-bot/commands'
 require 'slack-nfl-bot/app'
