+++ slack-nfl-bot/commands.rb
@@ -1,3 +1,4 @@
-require 'slack-nfl-bot/commands/nfl_bot'
+require 'slack-nfl-bot/commands/team'
+require 'slack-nfl-bot/commands/fact'
 require 'slack-nfl-bot/commands/ping'
 require 'slack-nfl-bot/commands/about'
