+++ slack-nfl-bot/commands/team.rb
@@ -1,30 +1,12 @@
 require "slack-ruby-bot"
-require "sqlite3"
 
 module SlackNFLBot
   module Commands
-    class NFLBot < SlackRubyBot::Commands::Base
+    class Team < SlackRubyBot::Commands::Base
 
       command(/(what(’|')s|which is) my team/, /(what|which) team((’|')s| is) mine/) do |client, data, match|
         team = get_team(data.user)
         client.say(text: "<@#{data.user}>, your team is the *#{team}*", channel: data.channel)
-      end
-
-      match /(what(’|')s| is) a fact about my team/ do |client, data, match|
-        team = get_team(data.user)
-        facts = find_facts(team)
-        say_fact(client, team, facts, data.channel) if facts
-      end
-
-      match /(what(’|')s| is) a fact about (the )?(.*)\?/ do |client, data, match|
-        if team = match[4]
-          facts = find_facts(team)
-          if facts
-            say_fact(client, team, facts, data.channel)
-          else
-            client.say(text: "I don't have any facts about the *#{team}* :cry:", channel: data.channel)
-          end
-        end
       end
 
       private
@@ -56,22 +38,8 @@ module SlackNFLBot
           rows.first
         end
 
-        def find_facts(team_name)
-          rows = database.execute <<-SQL
-            select fact from teams_facts tf join teams t
-              on tf.team_id = t.id
-              where lower(t.name) like lower('%#{team_name}%')
-          SQL
-
-          rows.first
-        end
-
-        def say_fact(client, team, facts, channel)
-          client.say(text: "Here's a fun fact about the *#{team}*: #{facts.sample}", channel: channel)
-        end
-
         def database
-          SQLite3::Database.new "db/production.db"
+          SlackNFLBot::Database.database
         end
       end
     end
