+++ slack-nfl-bot/database.rb
@@ -1,0 +1,9 @@
+require "sqlite3"
+
+module SlackNFLBot
+  class Database
+    def self.database
+      SQLite3::Database.new "db/#{ENV['RACK_ENV']}.db"
+    end
+  end
+end
